refactoring racing game and stringAddCalculator
.
