+++ src/main/java/calculator/string/adder/SpliterFactory.java
@@ -7,8 +7,7 @@ public class SpliterFactory {
     public static Spliter createSpliter(String s){
         Matcher m = Pattern.compile("//(.)N(.*)").matcher(s);
         if (m.find()) {
-            String customDelimiter = m.group(1);
-            return new CustomSpliter(customDelimiter);
+            return new CustomSpliter(m.group(1));
         }
         return new DefaultSpliter();
     }
