+++ src/main/java/racingcar/RacingGame.java
@@ -46,16 +46,4 @@ public class RacingGame {
     public static String getWinnerText(List<Car> winners) {
         return winners.stream().map((car) -> car.getName()).collect(Collectors.joining(","));
     }
-
-
-    public static void main(String[] args) {
-        String[] names = getNames();
-        int time = getTime();
-
-        RacingGame racingGame = new RacingGame(names);
-        for (int i = 0; i < time; i++) {
-            racingGame.moveAll();
-        }
-        racingGame.printResult();
-    }
 }
